pyproject.toml
[tool.poetry]
name = "wav2vec-speaker-identification"
version = "0.1.0"
description = ""
authors = ["Nik Vaessen <git@mail.vaessen.tech>"]
packages = [
    { include = "src" },
    { include = "tests" }
]

[tool.poetry.dependencies]
python = "^3.8"
tqdm = "^4.27"
scipy = "^1.6.0"
scikit-learn = "^0.24.0"
click = "^7.1.2"
pandas = "^1.2.1"
pytorch-lightning = "1.4.5"
pytorch-model-summary = "^0.1.2"
jupyter = "^1.0.0"
librosa = "^0.8.0"
matplotlib = "^3.3.4"
seaborn = "^0.11.1"
hydra-core = "^1.1.0"
psutil = "^5.8.0"
webdataset = "0.1.58"
yaspin = "2.0.0"
"hurry.filesize" = "^0.9"
python-dotenv = "^0.17.0"
torchaudio = "0.9.0"
speechbrain = "^0.5.5"
comet-ml = "^3.9.0"
lightning-bolts = "^0.3.3"
hydra-submitit-launcher = "^1.1.1"
wavaugment = "^0.2"
jupyterlab = "^3.0.14"
fairseq = "0.10.2"
jiwer = "^2.2.0"
datasets = "^1.8.0"
transformers = "^4.8.2"
hydra-optuna-sweeper = "^1.1.0"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
pytest = "^6.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
